Replace unsigned type by signed type
Losing one bit of size is no big deal in my use cases but having safe subtraction is nice.

+++ doublebeam-cpp/src/twopoint.cpp
@@ -204,7 +204,7 @@ slowness_t TwoPointRayTracing::trace(position_t source, position_t receiver,
                                 << "Receiver at " << stringify(receiver) << "  outside of model.");
     }
     // while the paper uses source based indexing, C++ doesn't.
-    auto source_index = model.layer_index(source_z).value() + 1;
+    auto source_index = static_cast<std::size_t>(model.layer_index(source_z).value() + 1);
     if (source_index > num_layers) {
         source_index = num_layers;
     }
